+++ example/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:pal/pal.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
 import 'router.dart';
 
 void main() {
@@ -15,7 +16,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Pal(
-      editorModeEnabled: true,
+      editorModeEnabled: false,
       appToken: APPLICATION_TOKEN, // don't forget to give us a token (create it from the web dashboard configuration tab)
       // --- your app is here --
       childApp: MaterialApp(
@@ -26,6 +27,13 @@ class MyApp extends StatelessWidget {
         title: 'Pal Demo',
         debugShowCheckedModeBanner: false,
         onGenerateRoute: route,
+        localizationsDelegates: [
+          GlobalWidgetsLocalizations.delegate,
+          GlobalMaterialLocalizations.delegate,
+        ],
+        supportedLocales: [
+          const Locale('fr')
+        ],
         theme: ThemeData(
           primarySwatch: Colors.blue,
           visualDensity: VisualDensity.adaptivePlatformDensity,