+++ lib/src/pal_navigator_observer.dart
@@ -31,10 +31,12 @@ class PalNavigatorObserver extends RouteObserver<PageRoute<dynamic>> implements 
   @override
   void didPush(Route<dynamic> route, Route<dynamic> previousRoute) {
     super.didPush(route, previousRoute);
-    if (route is PageRoute) {
-      _notify(route.settings);
-      _notifyRoute(route);
-    }
+      if (route is PageRoute) {
+        _notify(route.settings);
+        _notifyRoute(route);
+      }
+    // WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
+    // });
   }
 
   @override